Temp fix for mui and react-hook-form integration
